move fetch helpers, traverse through issues
