+++ app/main/api/importProjects.js
@@ -1,13 +1,14 @@
-import fetch from 'node-fetch';
-import { status, json } from '../../shared/helpers/fetch';
+import fetchPaginated from '../../shared/helpers/fetchPaginated';
+import { iterator, status, json, jsonAggregator } from '../../shared/helpers/fetch';
 
 export default function importProjects(accessToken, repoFullName) {
-  return fetch(`https://api.github.com/repos/${repoFullName}/issues?state=all`, {
+  return fetchPaginated(`https://api.github.com/repos/${repoFullName}/issues?state=all`, {
     method: 'GET',
     headers: {
       Authorization: `token ${accessToken}`,
     },
   })
-  .then(status)
-  .then(json);
+  .then(iterator(status))
+  .then(iterator(json))
+  .then(jsonAggregator);
 }
