+++ app/shared/helpers/fetch/index.js
@@ -1,0 +1,13 @@
+import iterator from './iterator';
+import json from './json';
+import jsonAggregator from './jsonAggregator';
+import status from './status';
+import text from './text';
+
+export {
+  iterator,
+  json,
+  jsonAggregator,
+  status,
+  text,
+};
