+++ app/shared/helpers/fetch/iterator.js
@@ -1,0 +1,5 @@
+export default function iterator(middleware) {
+  return responses => Promise.all(
+    responses.map(response => middleware(response))
+  );
+}
