+++ app/shared/helpers/fetch/json.js
@@ -1,0 +1,3 @@
+export default function json(response) {
+  return response.json();
+}
