+++ app/shared/helpers/fetch/jsonAggregator.js
@@ -1,0 +1,5 @@
+export default function jsonAggregator(responses) {
+  const aggregated = [].concat(...responses);
+
+  return aggregated;
+}
