+++ app/shared/helpers/fetch/text.js
@@ -1,0 +1,3 @@
+export default function text(response) {
+  return response.text();
+}
